mastermind

+++ src/main/java/com/example/projekt2/StageProperties.java
@@ -3,8 +3,8 @@ package com.example.projekt2;
 import javafx.scene.paint.Color;
 
 public class StageProperties {
-    public static int STAGE_WIDTH = 750;
-    public static int STAGE_HEIGHT = 550;
+    public static double STAGE_WIDTH = 750;
+    public static double STAGE_HEIGHT = 550;
     public static int FIELD_SIZE = 30;
     public static double RECTANGLE_SIZE = 45;
     public static Color[] COLORS = {
@@ -17,5 +17,4 @@ public class StageProperties {
             Color.BLUE,
             Color.PURPLE
     };
-
 }
